requirements.txt
altair==4.1.0
appdirs==1.4.3
argon2-cffi==20.1.0
asn1crypto==1.3.0
astor==0.8.1
astroid==2.3.3
async-generator==1.10
attrs==19.3.0
autobahn==21.3.1
Automat==20.2.0
azure-common==1.1.24
azure-storage-blob==2.1.0
azure-storage-common==2.1.0
backcall==0.2.0
base58==2.1.0
beautifulsoup4==4.9.1
bleach==3.1.5
blinker==1.4
bokeh==2.3.0
boto3==1.10.14
botocore==1.13.14
bottle==0.12.19
CacheControl==0.12.6
cachetools==4.2.1
cachy==0.3.0
certifi==2020.6.20
cffi==1.14.0
chardet==3.0.4
cleo==0.7.6
click==7.1.2
clikit==0.4.2
colorama==0.4.3
constantly==15.1.0
cryptography==3.4.7
cycler==0.10.0
Cython==0.29.17
dateparser==1.0.0
decorator==4.4.2
defusedxml==0.7.1
distlib==0.3.0
docutils==0.16
entrypoints==0.3
euroleague-scrapping==0.0.1.dev21
filelock==3.0.12
Flask==1.1.2
gitdb==4.0.7
GitPython==3.1.14
html5lib==1.0.1
hyperlink==21.0.0
idna==2.10
ijson==2.6.1
importlib-metadata==1.7.0
incremental==21.3.0
ipykernel==5.5.0
ipython==7.22.0
ipython-genutils==0.2.0
ipywidgets==7.6.3
isort==4.3.21
itsdangerous==1.1.0
jedi==0.18.0
jeepney==0.4.3
Jinja2==2.11.2
jmespath==0.9.4
joblib==1.0.1
jsonschema==3.2.0
jupyter==1.0.0
jupyter-client==6.1.12
jupyter-console==6.4.0
jupyter-core==4.7.1
jupyterlab-pygments==0.1.2
jupyterlab-widgets==1.0.0
keyring==21.2.1
kiwisolver==1.3.1
lazy-object-proxy==1.4.3
lockfile==0.12.2
lxml==4.6.3
MarkupSafe==1.1.1
matplotlib==3.3.3
mccabe==0.6.1
mistune==0.8.4
mpld3==0.5.2
msgpack==1.0.0
nbashots==0.1.0a9
nbclient==0.5.3
nbconvert==6.0.7
nbformat==5.1.2
nbstripout==0.3.9
nest-asyncio==1.5.1
notebook==6.3.0
numpy==1.17.4
oscrypto==1.2.0
packaging==20.4
pandas==0.25.3
pandocfilters==1.4.3
parso==0.8.1
pastel==0.2.0
patsy==0.5.1
pbr==5.4.5
pexpect==4.8.0
pickleshare==0.7.5
Pillow==8.1.0
pkginfo==1.5.0.1
pmdarima==1.8.0
poetry==1.0.5
prometheus-client==0.9.0
prompt-toolkit==3.0.18
protobuf==3.15.8
psycopg2==2.8.6
ptyprocess==0.6.0
pyarrow==3.0.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
pycryptodomex==3.9.7
pydeck==0.6.2
Pygments==2.5.2
PyJWT==1.7.1
pylev==1.3.0
pylint==2.4.3
pymssql==2.1.5
pyodbc==4.0.30
pyOpenSSL==19.1.0
pyparsing==2.4.7
pyrsistent==0.14.11
python-binance==0.7.10
python-dateutil==2.8.0
python-dotenv==0.17.0
pytz==2019.3
pywin32==300
pywin32-ctypes==0.2.0
pywinpty==0.5.7
PyYAML==5.4.1
pyzmq==22.0.3
qtconsole==5.0.3
QtPy==1.9.0
readme-renderer==26.0
regex==2021.4.4
requests==2.24.0
requests-toolbelt==0.9.1
rfc3986==1.4.0
s3transfer==0.2.1
scikit-learn==0.24.1
scipy==1.6.2
seaborn==0.6.0
SecretStorage==3.1.2
Send2Trash==1.5.0
service-identity==18.1.0
shellingham==1.3.2
six==1.15.0
smmap==4.0.0
snowflake==0.0.3
snowflake-connector-python==2.2.1
snowflake-sqlalchemy==1.2.1
soupsieve==1.9.6
SQLAlchemy==1.3.15
statsmodels==0.12.2
streamlit==0.80.0
terminado==0.9.3
testpath==0.4.4
threadpoolctl==2.1.0
toml==0.10.2
tomlkit==0.5.11
toolz==0.11.1
tornado==6.1
tqdm==4.48.0
traitlets==5.0.5
twine==1.15.0
Twisted==21.2.0
twisted-iocpsupport==1.0.1
txaio==21.2.1
typed-ast==1.4.0
typing-extensions==3.7.4.3
tzlocal==2.1
ujson==4.0.2
urllib3==1.25.9
validators==0.18.2
virtualenv==20.0.10
watchdog==2.0.2
wcwidth==0.2.5
webencodings==0.5.1
Werkzeug==1.0.1
widgetsnbextension==3.5.1
wrapt==1.11.2
zipp==3.1.0
zope.interface==5.4.0
